chore: Update Ubiquity installer to use rsync with additional options
This commit updates the customize.sh file to use rsync with the options -Aavx --update --delete when patching the Ubiquity installer. This will ensure that the /usr/share/ubiquity-slideshow/slides directory is synchronized with the /opt/slides directory, including deleting any files that are no longer present in /opt/slides. This change improves the patching process and ensures that the installer has the latest slideshow content.