chore: Update chroot_build.sh to install Fluent theme and gnome extensions
This commit updates the `chroot_build.sh` script to install the Fluent theme and gnome extensions. It removes the unnecessary `-i ubuntu` flag from the Fluent theme installation command and adds the installation command for the gnome extensions. These changes enhance the visual appearance and user experience of the system.