chore: Update upgrade.sh to install new kernel and packages
This commit updates the upgrade.sh script to install a new kernel and packages. The previous code only installed the linux-headers-generic-hwe-22.04 and linux-image-generic-hwe-22.04 packages, but now it also installs the linux-image-6.8.0-40-generic and linux-modules-extra-6.8.0-40-generic packages. This change ensures that the system is upgraded with the latest kernel and packages during the upgrade process.