chore: Add script to build all language modes
This commit adds a new script, build_all_langs.sh, which allows for building the application in multiple language modes. The script iterates over an array of language modes and language pack codes, updating the args.sh file with the corresponding values and running the build.sh script. This enhancement improves the localization support and allows for easier testing and deployment in different languages.