chore: Update grub.cfg for AnduinOS boot menu
This commit updates the grub.cfg file to modify the AnduinOS boot menu. The changes include updating the boot parameters for the "Try AnduinOS" and "Install AnduinOS" menu entries to remove the "verify-checksums" option. Additionally, the "OEM install (for manufacturers)" menu entry has been removed. These modifications improve the boot process and align it with the desired functionality of AnduinOS.